fix: Old versions of vyper do not play well with types

+++ contracts/Interfaces/curve/Curve.sol
@@ -44,8 +44,8 @@ interface ICurveFi {
         bool use_underlying
     ) external payable;
 
-    function coins(uint256) external returns (address);
-    function underlying_coins(uint256) external returns (address);
+    function coins(int128) external returns (address);
+    function underlying_coins(int128) external returns (address);
 
     function remove_liquidity_imbalance(uint256[2] calldata amounts, uint256 max_burn_amount) external;
